# nodejs/api/swagger.yaml
openapi: 3.0.0
info:
  title: Unified Data Repository Service API file for subscription data
  description: |
    Unified Data Repository Service (subscription data).
    The API version is defined in 3GPP TS 29.504.
    © 2019, 3GPP Organizational Partners (ARIB, ATIS, CCSA, ETSI, TSDSI, TTA, TTC).
    All rights reserved.
  version: '-'
...
servers:
- url: /
paths:
...
  /subscription-data/{ueId}/ue-update-confirmation-data/upu-data:
    get:
      tags:
      - Authentication UPU (Document)
      summary: Retrieves the UPU acknowledgement information of a UE
      operationId: QueryAuthUPU
      parameters:
      - name: ueId
        in: path
        description: UE id
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeId'
      - name: supportedFeatures
        in: query
        description: Supported Features
        required: false
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/SupportedFeatures'
      responses:
        200:
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpuData'
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: AuthenticationUPUDocument
    put:
      tags:
      - Authentication UPU (Document)
      summary: To store the UPU acknowledgement information of a UE
      operationId: CreateAuthenticationUPU
      parameters:
      - name: ueId
        in: path
        description: UE id
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeId'
      - name: supported-features
        in: query
        description: Supported Features
        required: false
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/SupportedFeatures'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpuData'
      responses:
        204:
          description: Expected response to a valid request
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: AuthenticationUPUDocument
...
  /subscription-data/{ueId}/operator-specific-data:
    get:
      tags:
      - Operator-Specific Data Container (Document)
      summary: Retrieves the operator specific data of a UE
      operationId: QueryOperSpecData
      parameters:
      - name: ueId
        in: path
        description: UE id
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeId'
      - name: fields
        in: query
        description: attributes to be retrieved
        required: false
        style: form
        explode: false
        schema:
          minItems: 1
          type: array
          items:
            type: string
      - name: supported-features
        in: query
        description: Supported Features
        required: false
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/SupportedFeatures'
      - name: If-None-Match
        in: header
        description: Validator for conditional requests, as described in RFC 7232,
          3.2
        required: false
        style: simple
        explode: false
        schema:
          type: string
      - name: If-Modified-Since
        in: header
        description: Validator for conditional requests, as described in RFC 7232,
          3.3
        required: false
        style: simple
        explode: false
        schema:
          type: string
      responses:
        200:
          description: Expected response to a valid request
          headers:
            Cache-Control:
              description: Cache-Control containing max-age, as described in RFC 7234,
                5.2
              style: simple
              explode: false
              schema:
                type: string
            ETag:
              description: Entity Tag, containing a strong validator, as described
                in RFC 7232, 2.3
              style: simple
              explode: false
              schema:
                type: string
            Last-Modified:
              description: Timestamp for last modification of the resource, as described
                in RFC 7232, 2.2
              style: simple
              explode: false
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperatorSpecificDataContainer'
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: OperatorSpecificDataContainerDocument
    patch:
      tags:
      - Operator-Specific Data Container (Document)
      summary: To modify operator specific data of a UE
      operationId: ModifyOperSpecData
      parameters:
      - name: ueId
        in: path
        description: UE id
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeId'
      requestBody:
        content:
          application/json-patch+json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/PatchItem'
        required: true
      responses:
        204:
          description: Expected response to a valid request
        403:
          description: modification is rejected
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: OperatorSpecificDataContainerDocument
...
  /subscription-data/{ueId}/context-data/ee-subscriptions:
    get:
      tags:
      - Event Exposure Subscriptions (Collection)
      summary: Retrieves the ee subscriptions of a UE
      operationId: Queryeesubscriptions
      parameters:
      - name: ueId
        in: path
        description: UE id
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeId'
      - name: supported-features
        in: query
        description: Supported Features
        required: false
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/SupportedFeatures'
      responses:
        200:
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EeSubscription'
                x-content-type: application/json
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: EventExposureSubscriptionsCollection
    post:
      tags:
      - Event Exposure Subscriptions (Collection)
      summary: Create individual EE subscription
      operationId: CreateEeSubscriptions
      parameters:
      - name: ueId
        in: path
        description: UE ID
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EeSubscription'
        required: true
      responses:
        201:
          description: Expected response to a valid request
          headers:
            Location:
              description: 'Contains the URI of the newly created resource, according
                to the structure: {apiRoot}/nudr-dr/<apiVersion>/subscription-data/{ueId}/context-data/ee-subscriptions/{subsId}'
              required: true
              style: simple
              explode: false
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EeSubscription'
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: EventExposureSubscriptionsCollection
  /subscription-data/{ueId}/context-data/ee-subscriptions/{subsId}:
    put:
      tags:
      - Event Exposure Subscription (Document)
      summary: Update an individual ee subscriptions of a UE
      operationId: UpdateEesubscriptions
      parameters:
      - name: ueId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeId'
      - name: subsId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EeSubscription'
      responses:
        204:
          description: Upon success, an empty response body shall be returned
        404:
          description: update of non-existing resource is rejected
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: EventExposureSubscriptionDocument
    delete:
      tags:
      - Event Exposure Subscription (Document)
      summary: Deletes a eeSubscription
      operationId: RemoveeeSubscriptions
      parameters:
      - name: ueId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeId'
      - name: subsId
        in: path
        description: Unique ID of the subscription to remove
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        204:
          description: Expected response to a successful subscription removal
      x-swagger-router-controller: EventExposureSubscriptionDocument
    patch:
      tags:
      - Event Exposure Subscription (Document)
      summary: Modify an individual ee subscription of a UE
      operationId: ModifyEesubscription
      parameters:
      - name: ueId
        in: path
        description: UE id
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeId'
      - name: subsId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json-patch+json:
            schema:
              minItems: 1
              type: array
              items:
                $ref: '#/components/schemas/PatchItem'
        required: true
      responses:
        204:
          description: Successful response
        403:
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        404:
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        default:
          description: Unexpected error
      x-swagger-router-controller: EventExposureSubscriptionDocument
...
  /subscription-data/group-data/{ueGroupId}/ee-subscriptions/{subsId}:
    put:
      tags:
      - Event Exposure Subscription (Document)
      summary: Update an individual ee subscription of a group of UEs or any UE
      operationId: UpdateEeGroupSubscriptions
      parameters:
      - name: ueGroupId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeGroupId'
      - name: subsId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EeSubscription'
      responses:
        204:
          description: Upon success, an empty response body shall be returned
        404:
          description: update of non-existing resource is rejected
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: EventExposureSubscriptionDocument
    delete:
      tags:
      - Event Exposure Subscription (Document)
      summary: Deletes a eeSubscription for a group of UEs or any UE
      operationId: RemoveEeGroupSubscriptions
      parameters:
      - name: ueGroupId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeGroupId'
      - name: subsId
        in: path
        description: Unique ID of the subscription to remove
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        204:
          description: Expected response to a successful subscription removal
      x-swagger-router-controller: EventExposureSubscriptionDocument
    patch:
      tags:
      - Event Exposure Subscription (Document)
      summary: Modify an individual ee subscription for a group of a UEs
      operationId: ModifyEeGroupSubscription
      parameters:
      - name: ueGroupId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeGroupId'
      - name: subsId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json-patch+json:
            schema:
              minItems: 1
              type: array
              items:
                $ref: '#/components/schemas/PatchItem'
        required: true
      responses:
        204:
          description: Successful response
        403:
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        404:
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        default:
          description: Unexpected error
      x-swagger-router-controller: EventExposureSubscriptionDocument
...
components:
  schemas:
...
    OperatorSpecificDataContainer:
      type: object
      properties:
        StringTypeElements:
          type: object
          additionalProperties:
            type: string
        IntegerTypeElements:
          type: object
          additionalProperties:
            type: integer
        NumberTypeElements:
          type: object
          additionalProperties:
            type: number
        BooleanTypeElements:
          type: object
          additionalProperties:
            type: boolean
...
    EeSubscription:
      required:
      - callbackReference
      - monitoringConfigurations
      type: object
      properties:
        callbackReference:
          $ref: '#/components/schemas/Uri'
        monitoringConfigurations:
          minProperties: 1
          type: object
          additionalProperties:
            $ref: '#/components/schemas/MonitoringConfiguration'
          description: A map (list of key-value pairs where ReferenceId serves as
            key) of MonitoringConfigurations
        reportingOptions:
          $ref: '#/components/schemas/ReportingOptions'
        supportedFeatures:
          $ref: '#/components/schemas/SupportedFeatures'
        subscriptionId:
          type: string
    VarUeGroupId:
      pattern: ^(extgroupid-[^@]+@[^@]+|anyUE)$
      type: string
...
    UpuData:
      required:
      - provisioningTime
      - ueUpdateStatus
      type: object
      properties:
        provisioningTime:
          $ref: '#/components/schemas/DateTime'
        ueUpdateStatus:
          $ref: '#/components/schemas/UeUpdateStatus'
        upuXmacIue:
          $ref: '#/components/schemas/UpuMac'
        upuMacIue:
          $ref: '#/components/schemas/UpuMac'
    UeUpdateStatus:
      type: string
      enum:
      - NOT_SENT
      - SENT_NO_ACK_REQUIRED
      - WAITING_FOR_ACK
      - ACK_RECEIVED
...
    VarUeId:
      pattern: ^(imsi-[0-9]{5,15}|nai-.+|msisdn-[0-9]{5,15}|extid-[^@]+@[^@]+|.+)$
      type: string
    SupportedFeatures:
      pattern: ^[A-Fa-f0-9]*$
      type: string
    ProblemDetails:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/Uri'
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          $ref: '#/components/schemas/Uri'
        cause:
          type: string
        invalidParams:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/InvalidParam'
        supportedFeatures:
          $ref: '#/components/schemas/SupportedFeatures'
    Uri:
      type: string
    InvalidParam:
      required:
      - param
      type: object
      properties:
        param:
          type: string
        reason:
          type: string
    PatchItem:
      required:
      - op
      - path
      type: object
      properties:
        op:
          $ref: '#/components/schemas/PatchOperation'
        path:
          type: string
        from:
          type: string
        value:
          nullable: true
    PatchOperation:
      anyOf:
      - type: string
        enum:
        - add
        - copy
        - move
        - remove
        - replace
        - test
      - type: string
...
    MonitoringConfiguration:
      required:
      - eventType
      type: object
      properties:
        eventType:
          $ref: '#/components/schemas/EventType'
        immediateFlag:
          type: boolean
        locationReportingConfiguration:
          $ref: '#/components/schemas/LocationReportingConfiguration'
    EventType:
      anyOf:
      - type: string
        enum:
        - LOSS_OF_CONNECTIVITY
        - UE_REACHABILITY_FOR_DATA
        - UE_REACHABILITY_FOR_SMS
        - LOCATION_REPORTING
        - CHANGE_OF_SUPI_PEI_ASSOCIATION
        - ROAMING_STATUS
        - COMMUNICATION_FAILURE
        - AVAILABILITY_AFTER_DNN_FAILURE
      - type: string
    LocationReportingConfiguration:
      required:
      - currentLocation
      type: object
      properties:
        currentLocation:
          type: boolean
        oneTime:
          type: boolean
        accuracy:
          $ref: '#/components/schemas/LocationAccuracy'
    LocationAccuracy:
      anyOf:
      - type: string
        enum:
        - CELL_LEVEL
        - TA_LEVEL
      - type: string
    ReportingOptions:
      type: object
      properties:
        maxNumOfReports:
          $ref: '#/components/schemas/MaxNumOfReports'
        expiry:
          $ref: '#/components/schemas/DateTime'
    MaxNumOfReports:
      type: integer
    DateTime:
      type: string
      format: date-time
...
    UpuMac:
      pattern: ^[A-Fa-f0-9]{32}$
      type: string
...